+++ peer/tworandomchoices/doc.go
@@ -18,7 +18,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package yarpc // import "go.uber.org/yarpc"
-
-// Version is the current version of YARPC.
-const Version = "1.33.0-dev"
+// Package tworandomchoices provides a load balancer implementation that picks
+// two peers at random and chooses the one with fewer pending requests.
+//
+// The Power of Two Choices in Randomized Load Balancing:
+// https://www.eecs.harvard.edu/~michaelm/postscripts/tpds2001.pdf
+//
+// The Power of Two Random Choices: A Survey of Techniques and Results:
+// https://www.eecs.harvard.edu/~michaelm/postscripts/handbook2001.pdf
+package tworandomchoices