+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- x/yarpctest: Add a retry option to HTTP/TChannel/GRPCRequest.
+- Added `peer/tworandomchoices`, an implementation of the Two Random Choices
+  load balancer algorithm.
+
 ## [1.32.4] - 2018-08-07
 ### Fixed
 - Address data races in yarpctest.FakeTransport
@@ -999,6 +1004,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[Unreleased]: https://github.com/yarpc/yarpc-go/compare/v1.32.4...HEAD
 [1.32.4]: https://github.com/yarpc/yarpc-go/compare/v1.32.3...v1.32.4
 [1.32.3]: https://github.com/yarpc/yarpc-go/compare/v1.32.2...v1.32.3
 [1.32.2]: https://github.com/yarpc/yarpc-go/compare/v1.32.1...v1.32.2